firstproject/requirements.txt
Django==2.2.2
django-angular==2.2
django-crispy-forms==1.7.2
django-filter==2.1.0
django-jenkins==0.110.0
djangorestframework==3.9.4
greenlet==0.4.15
html5lib==0.999999999
httplib2==0.9.2
Jinja2==2.10
jsonschema==2.6.0
psycopg2==2.8.3
Pygments==2.3.1
pytz==2019.1
requests==2.21.0
sqlparse==0.3.0
urllib3==1.24.2
virtualenv==16.5.0

